pyproject.toml
[project]
name = "sr-unstructured-adapter"
version = "0.1.0"
description = "Transform unstructured files into a unified JSON payload"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "SpiralReality"}]
dependencies = [
    "pydantic>=1.10,<3",
    "PyYAML>=6.0",
    "pdfminer.six>=20221105",
]

[project.scripts]
sr-adapt = "sr_adapter.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = ["-q"]
